app/requirements.txt
boto3==1.40.16
fastapi==0.116.1
h11==0.16.0
mangum==0.19.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings
PyJWT==2.10.1
pymongo==4.14.1
python-dateutil==2.9.0.post0
requests==2.32.5
uvicorn